Move SerializedDataField to coincidences app. The field is no longer used by histograms. Also update field reference in old histograms migration.
